+++ borax/calendars/festivals.py
@@ -1,36 +1,26 @@
 # coding=utf8
+import calendar
 import csv
 from collections import defaultdict
-from pathlib import Path
-import calendar
-import re
 from datetime import date
+from pathlib import Path
 from typing import Union, List, Iterator, Tuple, Optional
 
-from borax.calendars.lunardate import LunarDate, LCalendars, TERMS_CN
+from .lunardate import LunarDate, LCalendars, TERMS_CN
+from .store import (
+    Field, EncoderMixin, f_year, f_month, f_day, f_leap, f_index, f_reverse, f_schema
+)
 
 MDate = Union[date, LunarDate]
+FestivalCountdownIterable = Iterator[Tuple[int, List]]
 
 # Const values for date schema parameters
 YEAR_ANY = 0
 
 
-def date2mixed(date_obj):
-    if isinstance(date_obj, date):
-        return date_obj.strftime('0%Y%m%d0')
-    elif isinstance(date_obj, LunarDate):
-        return date_obj.strftime('1%y%A%B%l')
-    else:
-        raise TypeError('Unsupported type: {}'.format(date_obj.__class__.__name__))
-
-
-def mixed2date(src):
-    ds = DateSchemaFactory.from_string(src)
-    return ds.resolve()
-
-
-class DateSchema:
+class DateSchema(EncoderMixin):
     date_class = None
+    schema = None
 
     def __init__(self, *args, **kwargs):
         """Any subclass MUST call this initial in its own initial.
@@ -45,6 +35,18 @@ class DateSchema:
     @property
     def fuzzy(self):
         return self.year == YEAR_ANY
+
+    def __key(self):
+        return list(map(int, self.encode()))
+
+    def __hash__(self):
+        return hash(self.__key())
+
+    def __eq__(self, other):
+        return isinstance(self, type(other)) and self.__key() == other.__key()
+
+    def __str__(self):
+        return '<{}: {}>'.format(self.__class__.__name__, self.encode())
 
     # --------------------------  API  -----------------------------------
 
@@ -63,24 +65,35 @@ class DateSchema:
             date_obj = self.date_class.today()
         return self._countdown(self._normalize(date_obj))
 
-    def resolve(self, year=YEAR_ANY):
+    def resolve(self, year: int = YEAR_ANY) -> MDate:
+        """Return the date object in the solar / lunar year.
+        :param year:
+        :return:
+        """
         year = year or self.year
         if year:
             return self._resolve(year)
         raise ValueError('Unable resolve the date without a specified year.')
 
+    def resolve_solar(self, year: int) -> date:
+        """Return the date object in a solar year.
+        :param year:
+        :return:
+        """
+        if self.date_class == date:
+            return self.resolve(year)
+        else:
+            solar_date = LCalendars.cast_date(self.resolve(year), date)
+            offset = solar_date.year - year
+            if offset:
+                solar_date = LCalendars.cast_date(self.resolve(year - offset), date)
+            return solar_date
+
     # --------------------  Internal Methods  -------------------------------
 
     def _normalize(self, date_obj):
         date_class = self._get_date_class()
-        if not isinstance(date_obj, (date, LunarDate)):
-            raise TypeError('Unsupported type: {}'.format(date_obj.__class__.__name__))
-        if isinstance(date_obj, date_class):
-            return date_obj
-        if isinstance(date_class, date):
-            return date_obj.to_solar_date()
-        else:
-            return LunarDate.from_solar(date_obj)
+        return LCalendars.cast_date(date_obj, date_class)
 
     # Interfaces in the subclasses
 
@@ -101,30 +114,29 @@ class DateSchema:
 
 class SolarSchema(DateSchema):
     date_class = date
+    schema = 0
+    fields = [f_schema, f_year, f_month, f_day, f_reverse]
 
     def __init__(self, month, day, year=YEAR_ANY, reverse=0, **kwargs):
         self.year = year
         self.month = month
         self.day = day
-        self._reverse = reverse
-        if self._reverse == 1 and self.year == YEAR_ANY and self.month == 2:
-            raise ValueError('Unable resolve date for February without a specified year.')
+        self.reverse = reverse
         super().__init__(**kwargs)
-
-    def _check_day(self, year):
-        pass
 
     def _resolve(self, year):
 
-        if self._reverse == 0:
+        if self.reverse == 0:
             day = self.day
         else:
-            day = calendar.monthrange(self.year, self.month)[1] - self.day + 1
+            day = calendar.monthrange(year, self.month)[1] - self.day + 1
         return date(year, self.month, day)
 
 
 class LunarSchema(DateSchema):
     date_class = LunarDate
+    schema = 1
+    fields = [f_schema, f_year, f_month, f_day, f_leap]
 
     def __init__(self, month, day, year=YEAR_ANY, leap=0, ignore_leap=1, **kwargs):
         self.year = year
@@ -146,6 +158,8 @@ class LunarSchema(DateSchema):
 
 class WeekSchema(DateSchema):
     date_class = date
+    schema = 2
+    fields = [f_schema, f_year, f_month, f_index, Field(name='week', length=1)]
 
     def __init__(self, month, index, week, year=YEAR_ANY, **kwargs):
         self.year = year
@@ -160,13 +174,15 @@ class WeekSchema(DateSchema):
 
     @staticmethod
     def week_day(year: int, month: int, index: int, week: int) -> int:
-        i = 0
-        cal = calendar.Calendar()
-        for d, w in cal.itermonthdays2(year, month):
-            if d != 0 and w == week:
-                i += 1
-                if i == index:
-                    return d
+        w, ndays = calendar.monthrange(year, month)
+        if week >= w:
+            d0 = week - w + 1
+        else:
+            d0 = 8 - (w - week)
+        d = d0 + 7 * (index - 1)
+        if not (1 <= d <= ndays):
+            raise ValueError('Invalid day for this month.')
+        return d
 
     def __hash__(self):
         return hash((self.year, self.month, self.index, self.week))
@@ -174,6 +190,8 @@ class WeekSchema(DateSchema):
 
 class DayLunarSchema(DateSchema):
     date_class = LunarDate
+    schema = 3
+    fields = [f_schema, f_year, f_month, f_day, f_reverse]
 
     def __init__(self, month, day, year=YEAR_ANY, reverse=0, **kwargs):
         self.year = year
@@ -192,6 +210,8 @@ class DayLunarSchema(DateSchema):
 
 class TermSchema(DateSchema):
     date_class = date
+    schema = 4
+    fields = [f_schema, f_year, Field(name=None, length=2), f_index, Field(name=None, length=1)]
 
     def __init__(self, index, year=YEAR_ANY, **kwargs):
         self.year = year
@@ -203,32 +223,84 @@ class TermSchema(DateSchema):
 
 
 class DateSchemaFactory:
-    LOOKUP = [
-        # 公历
-        [re.compile(r'^(?P<schema>0)(?P<year>\d{4})(?P<month>\d{2})(?P<day>\d{2})(?P<reverse>[01])$'), SolarSchema],
-        # 农历
-        [re.compile(r'^(?P<schema>1)(?P<year>\d{4})(?P<month>\d{2})(?P<day>\d{2})(?P<leap>[01])$'), LunarSchema],
-        [re.compile(r'^(?P<schema>2)(?P<year>\d{4})(?P<month>\d{2})(?P<index>\d{2})(?P<week>[0-6])$'), WeekSchema],
-        [re.compile(r'^(?P<schema>3)(?P<year>\d{4})(?P<month>\d{2})(?P<day>\d{2})(?P<reverse>[01])$'), DayLunarSchema],
-        [re.compile(r'^(?P<schema>4)(?P<year>\d{4})00(?P<index>\d{2})0$'), TermSchema],
-    ]
+    schema_dict = {
+        0: SolarSchema,
+        1: LunarSchema,
+        2: WeekSchema,
+        3: DayLunarSchema,
+        4: TermSchema
+    }
 
-    @staticmethod
-    def from_string(raw: str, **kwargs) -> DateSchema:
-        for regex, cls in DateSchemaFactory.LOOKUP:
-            m = regex.match(raw)
-            if m:
-                kw = {k: int(v) for k, v in m.groupdict().items()}
-                return cls(**kw, **kwargs)
+    @classmethod
+    def from_string(cls, raw, **kwargs):
+        lg = len(raw)
+        if lg == 10:
+            short = False
+        elif lg == 6:
+            short = True
         else:
-            raise ValueError('Unable to match any schema for {}'.format(raw))
+            raise ValueError('Length expects 6 or 10, but {} got'.format(lg))
+        schema_code = int(raw[0])
+        schema_class = cls.schema_dict.get(schema_code)
+        schema = schema_class.decode(raw, short)
+        for k, v in kwargs.items():
+            setattr(schema, k, v)
+        return schema
+
+    @classmethod
+    def decode(cls, raw):
+        return cls.from_string(raw)
 
 
 # -------------------- Festival Dataset  ---------------------------------
-def read_dataset():
+LANG_FILES = {
+    'zh-Hans': 'FestivalData.txt'
+}
+
+
+class FestivalFactory:
+    """A container including a collection of festivals group by language or file.
+    """
+
+    def __init__(self, *, lang=None, file_path=None):
+        if lang:
+            file_path = Path(__file__).parent / LANG_FILES.get(lang)
+        if isinstance(file_path, str):
+            file_path = Path(file_path)
+        if file_path:
+            self._schema_list = list(read_dataset(file_path))
+        else:
+            self._schema_list = []
+
+    def add_festivals(self, festivals):
+        self._schema_list.extend(festivals)
+
+    def iter_festivals(self):
+        for festival in self._schema_list:
+            yield festival
+
+    def get_festival(self, name: str) -> DateSchema:
+        for schema in self._schema_list:
+            if schema.name == name:
+                return schema
+
+    def get_festival_names(self, date_obj: MDate) -> list:
+        return [schema.name for schema in self._schema_list if schema.match(date_obj)]
+
+    def iter_festival_countdown(self, countdown: Optional[int] = None,
+                                date_obj: MDate = None) -> FestivalCountdownIterable:
+        festival_names = defaultdict(list)
+        for schema in self._schema_list:
+            _offset = schema.countdown(date_obj)
+            if countdown is None or _offset <= countdown:
+                festival_names[_offset].append(schema.name)
+        for offset in sorted(festival_names.keys()):
+            yield offset, festival_names[offset]
+
+
+def read_dataset(file_path=None):
     field_names = ['src', 'name']
-    data_file = Path(__file__).parent / 'FestivalData.txt'
-    with data_file.open(encoding='utf8') as f:
+    with file_path.open(encoding='utf8') as f:
         reader = csv.DictReader(f, fieldnames=field_names)
         for row in reader:
             try:
@@ -237,10 +309,11 @@ def read_dataset():
                 continue
 
 
-def get_festival(name: str) -> DateSchema:
-    for schema in read_dataset():
-        if schema.name == name:
-            return schema
+# ---------- Shortcuts Method ----------
+
+def get_festival(name: str, lang: str = 'zh-Hans') -> DateSchema:
+    factory = FestivalFactory(lang=lang)
+    return factory.get_festival(name)
 
 
 def get_term(name: str) -> TermSchema:
@@ -248,13 +321,9 @@ def get_term(name: str) -> TermSchema:
     return TermSchema(index)
 
 
-def iter_festival_countdown(countdown: Optional[int] = None, date_obj: MDate = None) -> Iterator[Tuple[int, List]]:
+def iter_festival_countdown(countdown: Optional[int] = None, date_obj: MDate = None,
+                            lang: str = 'zh-Hans') -> FestivalCountdownIterable:
     """Return countdown of festivals.
     """
-    festival_names = defaultdict(list)
-    for schema in read_dataset():
-        _offset = schema.countdown(date_obj)
-        if countdown is None or _offset <= countdown:
-            festival_names[_offset].append(schema.name)
-    for offset in sorted(festival_names.keys()):
-        yield offset, festival_names[offset]
+    factory = FestivalFactory(lang=lang)
+    return factory.iter_festival_countdown(countdown, date_obj)